Add protocol definitions for Fetch API v10
This version is identical to v9, but allows for the compression type to be 'zstd', and allows
a new error code to be returned when fetching. We already had earlier code to inject a zstd
codec through the regular codec extension mechanism, this commit now simply makes it possible
to actually trigger that code.