Partial KIP-392 implementation
This commit adds enough code to make KafkaJS "understand" the "preferred replica" response that Kafka
2.4+ will include in fetch v11 responses, and use that preferred replica rather than the leader when
fetching.

This commit does not contain code to change the handling for some of the error cases, so we're basically
just doing retries/refreshes as before. The implementation for actually sending OffsetForLeaderEpoch requests
and handling the response is also not here, so as a result this is a "opportunistic" KIP-392 implementation
only: Iff Kafka sends us a hint, we will try to use it.

Note: This is based on work in https://github.com/figma/kafkajs/pull/1 by @jgao54