Throw errors when metadata is missing
Missing metadata means that something didn't correctly refresh it at a time it should
have, and not having the metadata will likely mean something else (such as the
`initializeConsumerOffsets` function) will crash.

We should rather crash out at the point where it is clear what is the problem, instead
of at a point where the only thing we know is "state is unexpected".