squash: Properly handle multiple clusters requesting a metadata refresh
The early attempt of simply combining topics derived from the known metadata
with the topics provided by the caller doesn't work if there are multiple callers
at the same time trying to each ask for _their_ topic to be added.

This commit then makes the function itself aware of "pending" topics that we're
currently in the process of getting the metadata for, and handles multiple callers
correctly.

Note that this is in theory also achievable through a big lock that only allows a
single mutation, similar to how the cluster manages its target topics -- the solution
presented here should scale better though when there are many requests coming quickly
as is probably common when starting up an app.